refactor: relocate node modules to agents
