--- src/core/nodes/__init__.py
@@ -1,3 +1,0 @@
-"""Async node handlers for the orchestration graph."""
-
-# TODO: Export concrete node implementations once stabilized.
